Added a grocery store list container that holds a list of all the grocery stores in the local area.
Added a loop that parses through a CSV and prints it to screen in a readable way.

Added TODO notes.

Created a public static list of string called GasStationList. Not sure about this one. Should I use the TextParser to generate the lists of stores and costs, or should I rely on static lists?