Added a new class (WebScraper) to scrape Wikipedia for the list of all known US gas stations.
The code was working to the point that it would do what I wanted, but once I got the list of stations in the variable during my debugging, I realized that I didn't need it to scrape anymore, and certainly not every time the program launches, and just added a gasStationList.txt to the solution.

I need to repeat this process with grocery stores and etc, and then populate all my lists with these .txt files.