Added an actual comarison engine to the ComparisonEngine method of the ComparisonMachine class. The algorithm works well, but the parsing system is passing USPS into gas station list.
May need to further parse data, choosing large chunks at a time and passing them to what is known. IE: Find all USPS and move to shippingCostList or something like that.

This entire program is working well other than the parsing. Things need to be cleaned up and documented better, but otherwise the workflow seems solid.

Going to upload this and do some time testing since the nested loop in the ComparisonEngine may be causing slowness.